+++ tests/conftest.py
@@ -1,0 +1,28 @@
+import pytest
+from chill_filter_web import create_app
+
+
+# from: https://flask.palletsprojects.com/en/stable/testing/
+
+@pytest.fixture()
+def app():
+    app = create_app()
+    app.config.update({
+        "TESTING": True,
+    })
+
+    # other setup can go here
+    yield app
+    
+    # clean up / reset resources here
+
+
+@pytest.fixture()
+def client(app):
+    return app.test_client()
+
+
+@pytest.fixture()
+def runner(app):
+    return app.test_cli_runner()
+
