+++ tests/test_basic.py
@@ -1,31 +1,3 @@
-import pytest
-from chill_filter_web import create_app
-
-
-# from: https://flask.palletsprojects.com/en/stable/testing/
-
-@pytest.fixture()
-def app():
-    app = create_app()
-    app.config.update({
-        "TESTING": True,
-    })
-
-    # other setup can go here
-    yield app
-    
-    # clean up / reset resources here
-
-
-@pytest.fixture()
-def client(app):
-    return app.test_client()
-
-
-@pytest.fixture()
-def runner(app):
-    return app.test_cli_runner()
-
 def test_front(client):
     response = client.get("/")
     print(response.data)
